added image

+++ AccountManagement/AccountManagement/Controllers/ProductController.cs
@@ -1,14 +1,18 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using AccountManagement.Contracts;
 using AccountManagement.Data;
+using AccountManagement.Data.DTO;
 using AccountManagement.Data.Model;
 using AccountManagement.Repository;
 using AutoMapper;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
+using static System.Net.Mime.MediaTypeNames;
 
 namespace AccountManagement.Controllers
 {
@@ -34,34 +38,68 @@ namespace AccountManagement.Controllers
 
 
         [HttpPost("Create")]
-        public IActionResult Create(ProductViewModel request)
+        public IActionResult Create(ProductCUDto request)
         {
             var product = _mapper.Map<Product>(request);
+            product.Category = _categoryRepository.FindById(request.CategoryId);
 
-            //product.Category = _categoryRepository.GetCategory(request.CategoryId);
-            // product.Category = _categoryRepository.GetCategory(request.CategoryId);
-
-            //   var category = categoryRepository.GetCategory(request.CategoryId);
-            //var cateogry
+            using (var ms = new MemoryStream())
+            {
+                request.Image.CopyTo(ms);
+                var fileBytes = ms.ToArray();
+            }
 
             var succeed = _productRepository.Create(product);
             return succeed ? Ok(new { Result = true }) : Ok(new { Result = false });
         }
 
 
-        [HttpPut("Update")]
-        public IActionResult Update(ProductViewModel request)
+        [HttpPut("Update/{id}")]
+        public IActionResult Update(int id, [FromForm] ImageDto file, ProductCUDto request)
         {
-            var product = _mapper.Map<Product>(request);
+            var product = _productRepository.FindById(id);
+            if (product == null) return NotFound("Product with id={id} does NOT exist");
+            product = _mapper.Map(request, product);
+
             var succeed = _productRepository.Update(product);
             return succeed ? Ok(new { Result = true }) : Ok(new { Result = false });
         }
 
-
-        [HttpDelete("Delete")]
-        public IActionResult Delete(ProductViewModel request)
+        [HttpGet("GetProduct/{id}")]
+        public IActionResult GetProduct(int id)
         {
-            var product = _mapper.Map<Product>(request);
+            var product = _productRepository.FindById(id);
+            if (product == null) return NotFound("Product does NOT exist");
+            return Ok(product);
+        }
+
+        [HttpPut("InsertImage/{id}")]
+        public IActionResult InsertImage(int id, [FromForm] ImageDto image)
+        {
+            var product = _productRepository.FindById(id);
+            using (var ms = new MemoryStream())
+            {
+                image.Image.CopyTo(ms);
+                var fileBytes = ms.ToArray();
+
+                if ((fileBytes.Length > 5e+6)) return BadRequest("Image size is to large , must be < 5mb ");
+                product.Image = fileBytes;
+
+
+            }
+
+            var succeed = _productRepository.Update(product);
+            return succeed ? Ok(new { Result = true }) : Ok(new { Result = false });
+
+        }
+
+
+        [HttpDelete("Delete/{id}")]
+        public IActionResult Delete(int id)
+        {
+            var product = _productRepository.FindById(id);
+            if (product == null) return NotFound();
+
             var succeed = _productRepository.Delete(product);
             return succeed ? Ok(new { Result = true }) : Ok(new { Result = false });
         }
@@ -81,10 +119,10 @@ namespace AccountManagement.Controllers
         }
 
 
-        [HttpGet("getallTEst")]
-        public async Task<IActionResult> GetCategorisAtProduct()
+        [HttpGet("GetProductsAndCategories")]
+        public async Task<IActionResult> GetProductsCategories()
         {
-            var products = await _productRepository.GetCategoryAtProducts();
+            var products = await _productRepository.GetProductsAndCategories();
             return Ok(products);
         }
 
@@ -94,6 +132,8 @@ namespace AccountManagement.Controllers
         //{
         //    var product=_productRepository.
         //}
+
+
     }
 }
